+++ packages/core/src/dtoMapper/models/responseDtoMapper.model.ts
@@ -1,4 +1,7 @@
-import { AnySchemaValidator } from '@forklaunch/validator';
+import {
+  AnySchemaValidator,
+  prettyPrintParseErrors
+} from '@forklaunch/validator';
 import { BaseEntity } from '../../database/mikro/models/entities/base.entity';
 import { DtoMapperConstructor } from '../interfaces/dtoMapper.interface';
 import { BaseDtoMapper, construct } from './baseDtoMapper.model';
@@ -43,7 +46,7 @@ export abstract class ResponseDtoMapper<
       this.dto
     );
     if (!parsedSchema.ok) {
-      throw new Error(`Invalid DTO: ${parsedSchema.error}`);
+      throw new Error(prettyPrintParseErrors(parsedSchema.errors, 'DTO'));
     }
     return this.dto;
   }