+++ packages/core/src/services/index.ts
@@ -1,4 +1,5 @@
 export * from './configInjector';
+export * from './getEnvVar';
 export * from './interfaces/baseService';
 export * from './types/configInjector.types';
 export * from './types/entityManager.types';