+++ packages/validator/index.ts
@@ -1,1 +1,2 @@
 export * from './src/shared/types/schema.types';
+export * from './src/shared/utils/prettyPrintParseErrors';