+++ packages/validator/tests/zod/schemaValidator.test.ts
@@ -1,6 +1,7 @@
 import { generateSchema } from '@anatine/zod-openapi';
 import { ZodLiteral, ZodObject, ZodTypeAny, z } from 'zod';
 import { UnboxedObjectSchema } from '../../src/shared/types/schema.types';
+import { prettyPrintParseErrors } from '../../src/shared/utils/prettyPrintParseErrors';
 import {
   array,
   enum_,
@@ -251,7 +252,7 @@ describe('zod schema validator tests', () => {
     });
     expect(failedParse.ok).toBe(false);
     if (!failedParse.ok) {
-      expect(failedParse.error)
+      expect(prettyPrintParseErrors(failedParse.errors))
         .toBe(`Validation failed with the following errors:
 1. Path: hello > world
    Message: Expected string, received number